[WOE] Implement Virtue of Knowledge
